pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "arst"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scikit-image>=0.21",
    "opencv-python>=4.8",
]

[tool.setuptools.packages.find]
where = ["src"]
